[Feat] Hexagoanl Timer
