+++ GongdeunTop/GongdeunTop/Views/Shape/CubeHexagon.swift
@@ -1,0 +1,66 @@
+//
+//  CubeHexagon.swift
+//  GongdeunTop
+//
+//  Created by Martin on 2023/04/18.
+//
+
+import SwiftUI
+
+struct CubeHexagon: Shape {
+    var radius: CGFloat
+
+    init(radius: CGFloat) {
+        self.radius = radius
+    }
+    
+    let points: [CGPoint] = [.init(x: sqrt(3 / 4), y: -0.5),
+                             .init(x: 0, y: -1),
+                             .init(x: -sqrt(3 / 4), y: -0.5),
+                             .init(x: -sqrt(3 / 4), y: 0.5),
+                             .init(x: 0, y: 1),
+                             .init(x: sqrt(3 / 4), y: 0.5)]
+     
+    func path(in rect: CGRect) -> Path {
+        let originX: CGFloat = rect.width / 2
+        let originY: CGFloat = rect.height / 2
+        
+        var path = Path()
+        
+        path.move(to: CGPoint(x: radius, y: 0))
+        for index in 0..<points.count {
+            let x = points[index].x * radius
+            let y = points[index].y * radius
+            
+            if index == 0 {
+                path.move(to: CGPoint(x: x, y: y))
+            } else {
+                path.addLine(to: CGPoint(x: x, y: y))
+            }
+            
+            if index % 2 == 0 {
+                path.addLine(to: CGPoint(x: 0, y: 0))
+                path.addLine(to: CGPoint(x: x, y: y))
+            }
+        }
+
+        path.closeSubpath()
+        
+        
+        return path.offsetBy(dx: originX, dy: originY)
+    }
+}
+
+
+struct MyShape: View {
+    var body: some View {
+        CubeHexagon(radius: 100)
+            .stroke()
+    }
+}
+
+struct MyShape_Previews: PreviewProvider {
+    static var previews: some View {
+        MyShape()
+    }
+}
